// src/app/shared/interfaces/sales-filter.ts
export interface ShopFilter {
    status?: string;
    productStage?: string;
    pageNumber?: string;
    pageSize?: string;
    thicknessMin?: string;
    thicknessMax?: string;
    widthMin?: string;
    widthMax?: string;
    lengthMin?: string;
    lengthMax?: string;
    productCategory?: string;
    productClass?: string;
    productShape?: string;
}